pdf refactoring
Continuing with the refactoring of the other parts of the code, I've
refactored some of the initial functions involved with computing/evaluating
PDFs to the `pdf` module, which is now loaded in on initialization.

As always, the `frankenz` (internal) module is still a mess and a work
in progress and is not loaded in.