redshift inference v4
A final set of updates. This involves a few big changes other than cleaning up
content and notation.

The biggest is that I've now included a physical
redshift prior for the population inference stage. This significantly
improves the results for both the redshift grid and the redshift samples. This
involved integrating a Heavyside function over the samples. I think we can
probably use a similar prior when applying our results to photometry subject
to selection boundaries.

I've also compared behavior with a 2-level hierarchical model where I allow
the alphas to float. While this improved mixing, it added a significant
burden to the computation time since we have to use HMC to sample the
conditional distribution over the alphas. It might be possible to use
a conjugate prior in the future, but I think this at least demonstrates that
the default 1-level Dirichlet model seems to work well enough for our
purposes.

The code, while straightforward, is still a bit of a mess. I need to implement
these samplers internally to allow for better sampling. Also, it appears that
there are some mixing issues with the 1-level Dirichlet Hierarchical model;
these can probably be alleviated with some simulated annealing/parallel
tempering, so I should double check if I can implement that.