photometric inference v1
Continuing the basic refactoring project as I go through the notebooks.
Much of this is a bit rougher than I'd like but at least the general
structure of the notebooks is cleaner. At some future point I will add a lot
of the fitting, etc. so that it happens internally via some wrapper to
avoid all the long `for` loops and related things.

Big changes this time around:
- Slight tweak to the mock data to properly store the model redshift grid.
- Notebook 3.1 has been completely redone. It looks much nicer now.
- Continued refactoring `frankenz.py` into other subpackages. Added
`plotting.py` and `nn.py`.
- The loglikelihood functions have now all been wrapped within one
central function `loglike`. I've also added the capability to include
errors in the color-based likelihood fits.
- I've labeled all other notebooks [OLD] where appropriate prior to
talking about this work at Amherst.