photometry demo
This commit adds back the photometry demo with SDSS along with the application
to mock HSC data.

I've also reorganized things a bit and have changed `loglike` so that
it always returns to "raw" chi2 value as well as the log-likelihood. Those
are more useful to remove objects with terrible fits and are somewhat more
intuitive to work with.