fitting refactoring
This commit refactors a bunch of the fitting methodology so that it mirrors
a simple `sklearn`-style format. Fitting objects now will contain:
- `__init__`: loads in the models and initializes any relevant data structures
- `fit`: takes in data plus a lprob function and computes the ln(prob)'s.
- `predict`: takes in model labels and computes the 1-D target PDFs.
- `fit_predict`: combines `fit` and `predict` so that intermediate fits
don't have to be stored.

As part of this, a number of other methods have been tweaked to allow them
to be easier to pass/specify during various stages of fitting. This mostly
involves adding `*args` and `**kwargs` options for passing things.

There was also some issues with Windows touching all my files a while back,
which I just fixed by converting everything from DOS to Unix format. This
just meant everything got touched again, leading to supposedly huge commits.

Finally, I split up some of the fitting. The "sparsity" stuff will now be in
a separate notebook since this one was getting quite long.
This makes sense in retrospect, because that's really a separate
concept, and dealing with it will necessarily entail covering a bunch of
other stuff now that I'm trying to integrate in the SOM and GNG.

This commit resolves #8 by implementing a new `BruteForce` object. I'll also
claim it resolves #9 since it introduces a new workflow that I will try to
emulate going forward. Hopefully after some changes to the nearest-neighbor
searches I'll be able to deal with #5 and #10 as well.