dim_prior changes
After working through the math, I think the actual trans-dimensional
likelihood is `Gamma(chisq|a=N/2)`, where `Gamma(x|a)` is the Gamma
*distribution* with shape parameter `a` and `chisq` is the
computed chi-square value. This is the result of assuming that
the observation `g` and model `h` are actually noisy realizations of the
same underlying galaxy `i`, which implies that the distribution of the
normalized residuals in the multivariate normal pdf is actually chi-square.
Exploiting conjugacy of the Gamma and Normal distributions (since chi-square
is a special case of Gamma) then immediately implies the relevant result.

I've now implemented this in the default log-likelihood functions, along
with some small reorganization.