scale_err, node_pdf bugfix
Added in support for computing and passing back errors on the scale factors,
since computing them is trivial given they are already involved in
computing the maximum-likelihood scale factor. This involved changing up
some parts of the interface both externally (since this is now passed back)
and internally (since each of the fitting objects needs to keep track of
the errors as well, if needed).

In addition, I also fixed some bugs involved with computing the PDFs
from just fitting the nodes of the network (`nodes_only=True`). This now
works properly, and I've added in some comparisons in the 3rd notebook.