gng support
This commit resolves #11 by adding in support for the Growing Neural Gas
algorithm. My implementation is very similar to that from the original one
proposed by Fritzke (1995), except that the network updates (adding vertices,
pruning edges) happen at the same time to avoid querying the network overly
often. The implementation is done through the `networkx` package, which is
really nice, and is partially based off https://github.com/ansrivas/GNG.

Minor changes:
- Added in support for keeping track of `scale_errs` within `fitting` objects
- Fixed a few SOM-related bugs.

Next on the agenda: implementing a deprojection module for the GNG and getting
the plotting utilities for the SOM and GNG up and running.