v0.1.0
This commit comes close to being the first "stable" version of the code,
complete with most of the baseline plotting utilities and functionality.
A summary of the changes is below:

- `fitting` objects now default to using `logprob`, which is implemented
  in `frankenz.pdf` and just wraps `loglike`. This simplifies some of the
  internal code and makes it easier to specify various configurations
  without passing custom user-defined functions.

- Some small bugfixes associated with fitting.

- Added in/improved plotting functions for examining PDFs.

- Added in functions to resample PDFs and to generate a whole slew of summary
  statistics. This should now give **exactly** the same results as the
  HSC photo-z pipeline.

- Modified defaults for networks so that they train in color space rather
  than flux space. This should hopefully lead to more stable behavior.

- Split up photometry demos to focus on fitting vs interpretation/plotting.

Only thing that's left is to get the network plots ported over/improved
from the old code, and we should be in business! If everything works out
tomorrow, I'll push `frankenz` to `pypi` to make it easy to get `brutus`
[name pending] set up.