v0.1.5
This commit finally resolves #13 by adding in plotting utilities for networks.
These include `plot2d_network`, which plots quantities in 2-D, and `plot_node`,
which plots the SED in a given node along with the relevant (scaled)
distribution of the objects associated with it.

Small bugfixes:
- SOM and GNG now properly track scales. Default kwargs have been updated
  to speed up fitting by ignoring model errors.
- Fixed an issue with `get_node` that was preventing it from searching out
  neighbors properly.
- Fixed a small bug that meant that PDFs were being returned slightly
  unnormalized. This wasn't a big deal for most things, but could throw
  off some parts of the code.

Last by not least -- `frankenz` is now on PyPI! Hooray!