v0.2.0
This commit involves some larger changes and some smaller bugfixes.

The biggest change is that `discrete` is now mostly a post-processing option
for all `network`-style objects rather than one that gets declared upon
initially training. While its not the optimal structure for pure performance,
its designed mostly for ease of integration into other parts of the code.
`discrete` was mostly intended to be a comparison anyway, so these changes
shouldn't be too offensive.

I've implemented additional safety checks in `loglike` to ensure bad data
is automatically flagged in case the user misses some of it. I should likely
port this over to some other codes I'm writing as well.

The remainder of this commit mostly includes a couple small bugfixes. As such,
this version should now be stable enough to comprise the "core" functionality
of the code (hence the version change).