v0.3.0
This commit cleans up a bunch of notation and derivations in earlier notebooks
and includes a new notebook outlining population inference. To accompany this,
I've now added a new `samplers` import which will contain samplers that
will be able to apply various models to predict redshift populations from
PDFs. The first addition is the `population_sampler`, which samples from
the 0th-level model where you just find the N(z) that is maximally consistent
with the data.